+++ src/loader/js/yui3.js
@@ -1083,6 +1083,12 @@ Y.mix(YUI.Env[Y.version].modules, {
         ]
     },
     "datatype-date-parse": {},
+    "datatype-list-format": {
+        "lang": [],
+        "requires": [
+            "yui-base"
+        ]
+    },
     "datatype-number": {
         "use": [
             "datatype-number-parse",
@@ -1090,6 +1096,27 @@ Y.mix(YUI.Env[Y.version].modules, {
         ]
     },
     "datatype-number-format": {},
+    "datatype-number-format-advanced": {
+        "lang": [],
+        "requires": [
+            "intl-common",
+            "datatype-number-format",
+            "datatype-number-parse"
+        ]
+    },
+    "datatype-number-format-ecma": {
+        "condition": {
+            "name": "datatype-number-format-ecma",
+            "test": function (Y) {
+	return (typeof window !== 'undefined' && window.Intl !== undefined);
+},
+            "trigger": "datatype-number-format-advanced",
+            "when": "after"
+        },
+        "requires": [
+            "intl"
+        ]
+    },
     "datatype-number-parse": {},
     "datatype-xml": {
         "use": [
@@ -1761,6 +1788,11 @@ Y.mix(YUI.Env[Y.version].modules, {
     "intl-base": {
         "requires": [
             "yui-base"
+        ]
+    },
+    "intl-common": {
+        "requires": [
+            "intl"
         ]
     },
     "io": {
@@ -2946,4 +2978,4 @@ Y.mix(YUI.Env[Y.version].modules, {
         ]
     }
 });
-YUI.Env[Y.version].md5 = '9cfdc7e62873313018cc48b86f2a53ca';
+YUI.Env[Y.version].md5 = '3b9d283e601f70d19bf722d7b26cf13c';