+++ src/number/meta/ecma.js
@@ -1,0 +1,3 @@
+function (Y) {
+	return (typeof window !== 'undefined' && window.Intl !== undefined);
+}