+++ src/simpleyui/js/concat.js
@@ -5479,14 +5479,23 @@ add('load', '1', {
 },
     "trigger": "autocomplete-list"
 });
-// dd-gestures
+// datatype-number-format-ecma
 add('load', '2', {
+    "name": "datatype-number-format-ecma",
+    "test": function (Y) {
+	return (typeof window !== 'undefined' && window.Intl !== undefined);
+},
+    "trigger": "datatype-number-format-advanced",
+    "when": "after"
+});
+// dd-gestures
+add('load', '3', {
     "name": "dd-gestures",
     "trigger": "dd-drag",
     "ua": "touchEnabled"
 });
 // dom-style-ie
-add('load', '3', {
+add('load', '4', {
     "name": "dom-style-ie",
     "test": function (Y) {
 
@@ -5517,14 +5526,14 @@ add('load', '3', {
     "trigger": "dom-style"
 });
 // editor-para-ie
-add('load', '4', {
+add('load', '5', {
     "name": "editor-para-ie",
     "trigger": "editor-para",
     "ua": "ie",
     "when": "instead"
 });
 // event-base-ie
-add('load', '5', {
+add('load', '6', {
     "name": "event-base-ie",
     "test": function(Y) {
     var imp = Y.config.doc && Y.config.doc.implementation;
@@ -5533,7 +5542,7 @@ add('load', '5', {
     "trigger": "node-base"
 });
 // graphics-canvas
-add('load', '6', {
+add('load', '7', {
     "name": "graphics-canvas",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -5545,7 +5554,7 @@ add('load', '6', {
     "trigger": "graphics"
 });
 // graphics-canvas-default
-add('load', '7', {
+add('load', '8', {
     "name": "graphics-canvas-default",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -5557,7 +5566,7 @@ add('load', '7', {
     "trigger": "graphics"
 });
 // graphics-svg
-add('load', '8', {
+add('load', '9', {
     "name": "graphics-svg",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -5570,7 +5579,7 @@ add('load', '8', {
     "trigger": "graphics"
 });
 // graphics-svg-default
-add('load', '9', {
+add('load', '10', {
     "name": "graphics-svg-default",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -5583,7 +5592,7 @@ add('load', '9', {
     "trigger": "graphics"
 });
 // graphics-vml
-add('load', '10', {
+add('load', '11', {
     "name": "graphics-vml",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -5593,7 +5602,7 @@ add('load', '10', {
     "trigger": "graphics"
 });
 // graphics-vml-default
-add('load', '11', {
+add('load', '12', {
     "name": "graphics-vml-default",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -5603,7 +5612,7 @@ add('load', '11', {
     "trigger": "graphics"
 });
 // history-hash-ie
-add('load', '12', {
+add('load', '13', {
     "name": "history-hash-ie",
     "test": function (Y) {
     var docMode = Y.config.doc && Y.config.doc.documentMode;
@@ -5614,13 +5623,13 @@ add('load', '12', {
     "trigger": "history-hash"
 });
 // io-nodejs
-add('load', '13', {
+add('load', '14', {
     "name": "io-nodejs",
     "trigger": "io-base",
     "ua": "nodejs"
 });
 // json-parse-shim
-add('load', '14', {
+add('load', '15', {
     "name": "json-parse-shim",
     "test": function (Y) {
     var _JSON = Y.config.global.JSON,
@@ -5647,7 +5656,7 @@ add('load', '14', {
     "trigger": "json-parse"
 });
 // json-stringify-shim
-add('load', '15', {
+add('load', '16', {
     "name": "json-stringify-shim",
     "test": function (Y) {
     var _JSON = Y.config.global.JSON,
@@ -5670,13 +5679,13 @@ add('load', '15', {
     "trigger": "json-stringify"
 });
 // scrollview-base-ie
-add('load', '16', {
+add('load', '17', {
     "name": "scrollview-base-ie",
     "trigger": "scrollview-base",
     "ua": "ie"
 });
 // selector-css2
-add('load', '17', {
+add('load', '18', {
     "name": "selector-css2",
     "test": function (Y) {
     var DOCUMENT = Y.config.doc,
@@ -5687,7 +5696,7 @@ add('load', '17', {
     "trigger": "selector"
 });
 // transition-timer
-add('load', '18', {
+add('load', '19', {
     "name": "transition-timer",
     "test": function (Y) {
     var DOCUMENT = Y.config.doc,
@@ -5703,13 +5712,13 @@ add('load', '18', {
     "trigger": "transition"
 });
 // widget-base-ie
-add('load', '19', {
+add('load', '20', {
     "name": "widget-base-ie",
     "trigger": "widget-base",
     "ua": "ie"
 });
 // yql-jsonp
-add('load', '20', {
+add('load', '21', {
     "name": "yql-jsonp",
     "test": function (Y) {
     /* Only load the JSONP module when not in nodejs or winjs
@@ -5721,14 +5730,14 @@ add('load', '20', {
     "when": "after"
 });
 // yql-nodejs
-add('load', '21', {
+add('load', '22', {
     "name": "yql-nodejs",
     "trigger": "yql",
     "ua": "nodejs",
     "when": "after"
 });
 // yql-winjs
-add('load', '22', {
+add('load', '23', {
     "name": "yql-winjs",
     "trigger": "yql",
     "ua": "winjs",
@@ -6607,14 +6616,23 @@ add('load', '1', {
 },
     "trigger": "autocomplete-list"
 });
-// dd-gestures
+// datatype-number-format-ecma
 add('load', '2', {
+    "name": "datatype-number-format-ecma",
+    "test": function (Y) {
+	return (typeof window !== 'undefined' && window.Intl !== undefined);
+},
+    "trigger": "datatype-number-format-advanced",
+    "when": "after"
+});
+// dd-gestures
+add('load', '3', {
     "name": "dd-gestures",
     "trigger": "dd-drag",
     "ua": "touchEnabled"
 });
 // dom-style-ie
-add('load', '3', {
+add('load', '4', {
     "name": "dom-style-ie",
     "test": function (Y) {
 
@@ -6645,14 +6663,14 @@ add('load', '3', {
     "trigger": "dom-style"
 });
 // editor-para-ie
-add('load', '4', {
+add('load', '5', {
     "name": "editor-para-ie",
     "trigger": "editor-para",
     "ua": "ie",
     "when": "instead"
 });
 // event-base-ie
-add('load', '5', {
+add('load', '6', {
     "name": "event-base-ie",
     "test": function(Y) {
     var imp = Y.config.doc && Y.config.doc.implementation;
@@ -6661,7 +6679,7 @@ add('load', '5', {
     "trigger": "node-base"
 });
 // graphics-canvas
-add('load', '6', {
+add('load', '7', {
     "name": "graphics-canvas",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -6673,7 +6691,7 @@ add('load', '6', {
     "trigger": "graphics"
 });
 // graphics-canvas-default
-add('load', '7', {
+add('load', '8', {
     "name": "graphics-canvas-default",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -6685,7 +6703,7 @@ add('load', '7', {
     "trigger": "graphics"
 });
 // graphics-svg
-add('load', '8', {
+add('load', '9', {
     "name": "graphics-svg",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -6698,7 +6716,7 @@ add('load', '8', {
     "trigger": "graphics"
 });
 // graphics-svg-default
-add('load', '9', {
+add('load', '10', {
     "name": "graphics-svg-default",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -6711,7 +6729,7 @@ add('load', '9', {
     "trigger": "graphics"
 });
 // graphics-vml
-add('load', '10', {
+add('load', '11', {
     "name": "graphics-vml",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -6721,7 +6739,7 @@ add('load', '10', {
     "trigger": "graphics"
 });
 // graphics-vml-default
-add('load', '11', {
+add('load', '12', {
     "name": "graphics-vml-default",
     "test": function(Y) {
     var DOCUMENT = Y.config.doc,
@@ -6731,7 +6749,7 @@ add('load', '11', {
     "trigger": "graphics"
 });
 // history-hash-ie
-add('load', '12', {
+add('load', '13', {
     "name": "history-hash-ie",
     "test": function (Y) {
     var docMode = Y.config.doc && Y.config.doc.documentMode;
@@ -6742,13 +6760,13 @@ add('load', '12', {
     "trigger": "history-hash"
 });
 // io-nodejs
-add('load', '13', {
+add('load', '14', {
     "name": "io-nodejs",
     "trigger": "io-base",
     "ua": "nodejs"
 });
 // json-parse-shim
-add('load', '14', {
+add('load', '15', {
     "name": "json-parse-shim",
     "test": function (Y) {
     var _JSON = Y.config.global.JSON,
@@ -6775,7 +6793,7 @@ add('load', '14', {
     "trigger": "json-parse"
 });
 // json-stringify-shim
-add('load', '15', {
+add('load', '16', {
     "name": "json-stringify-shim",
     "test": function (Y) {
     var _JSON = Y.config.global.JSON,
@@ -6798,13 +6816,13 @@ add('load', '15', {
     "trigger": "json-stringify"
 });
 // scrollview-base-ie
-add('load', '16', {
+add('load', '17', {
     "name": "scrollview-base-ie",
     "trigger": "scrollview-base",
     "ua": "ie"
 });
 // selector-css2
-add('load', '17', {
+add('load', '18', {
     "name": "selector-css2",
     "test": function (Y) {
     var DOCUMENT = Y.config.doc,
@@ -6815,7 +6833,7 @@ add('load', '17', {
     "trigger": "selector"
 });
 // transition-timer
-add('load', '18', {
+add('load', '19', {
     "name": "transition-timer",
     "test": function (Y) {
     var DOCUMENT = Y.config.doc,
@@ -6831,13 +6849,13 @@ add('load', '18', {
     "trigger": "transition"
 });
 // widget-base-ie
-add('load', '19', {
+add('load', '20', {
     "name": "widget-base-ie",
     "trigger": "widget-base",
     "ua": "ie"
 });
 // yql-jsonp
-add('load', '20', {
+add('load', '21', {
     "name": "yql-jsonp",
     "test": function (Y) {
     /* Only load the JSONP module when not in nodejs or winjs
@@ -6849,14 +6867,14 @@ add('load', '20', {
     "when": "after"
 });
 // yql-nodejs
-add('load', '21', {
+add('load', '22', {
     "name": "yql-nodejs",
     "trigger": "yql",
     "ua": "nodejs",
     "when": "after"
 });
 // yql-winjs
-add('load', '22', {
+add('load', '23', {
     "name": "yql-winjs",
     "trigger": "yql",
     "ua": "winjs",